feat: add combo to cart, view cart

+++ src/main/java/com/nutrimeal/nutrimeal/model/OrderDetail.java
@@ -21,7 +21,7 @@ public class OrderDetail {
     private int orderDetailQuantity;
 
     @Column(name = "order_detail_price")
-    private float orderDetailPrice;
+    private int orderDetailPrice;
 
     @ManyToOne
     @JoinColumn(name = "combo_id")
